perf: ProfileCardContainer 컴포넌트 분리 및 memo로 메모이제이션 수행
